Refactor project structure, add Vite support, and optimize Telegram Web App
