+++ index.html
@@ -9,6 +9,7 @@
     <meta name="format-detection" content="telephone=no" />
     <title>ASTROT - Космическая Астрология с Котеусом</title>
     <script src="https://telegram.org/js/telegram-web-app.js"></script>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/framework7@8.0.0/css/framework7-bundle.css">
   </head>
   <body>
     <div id="root"></div>
