+++ src/App.tsx
@@ -1,12 +1,11 @@
 import React, { useEffect, useState } from 'react';
 import { App as KonstaApp } from 'konsta/react';
 import { App as F7App, View } from 'framework7-react';
-import routes from './routes.js';
+import routes from './routes.tsx';
 import TelegramWrapper from './components/TelegramWrapper';
 import ErrorBoundary from './components/ErrorBoundary';
 import Framework7 from 'framework7/lite-bundle';
 import Framework7React from 'framework7-react';
-import 'framework7/framework7-bundle.css';
 
 // Register Framework7 React plugin (required for Framework7 components to work correctly)
 Framework7.use(Framework7React);
